feat: implement the ability to create orders
This adds custom validation to the existing model to ensure we don't
create invalid orders, migrations will be implemented for data integrity
on the server also.

A model helper was implemented to allow easy processing of the errors
from Sequelize.

This also includes tests for the helper and the facade.